pyproject.toml
[project]
name = "drf-social-network"
version = "0.1.0"
description = ""
authors = [
    {name = "usman",email = "usmanakkaev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "djangorestframework (>=3.15.2,<4.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
